Add tests + clean code

+++ rs-microservice/regards-modules/session/session-management/session-management-dao/src/main/java/fr/cnes/regards/framework/modules/session/management/dao/ISourceManagerRepository.java
@@ -38,7 +38,7 @@ public interface ISourceManagerRepository extends JpaRepository<Source, Long>, J
 
     Optional<Source> findByName(String name);
 
-    void deleteByNbSessions(int noSession);
+    void deleteByNbSessions(long noSession);
 
     void deleteByName(String source);
